Template, Aliases, LIMIT array and new dics.
Added template for examples.
Added options to create aliases for tables and fields.
Started on improved documentation.
Switched LIMIT clause from two seperate keys into one array.
Added 'IS NULL' and 'IS NOT NULL' options to the WHERE clause. New
whereElementNull() function to accomodate this.
Added missing backticks around tables and fields in the SQL.
Added a print_r of the query to the showData() function.