Version 2
Full rewrite of the class.

Added the option to ive tables and column alias's.

Removed the tableAction() function and implemented this directly into
the buildTables() function.

Rebuild of the main CRUD() function to remove duplicated steps and
optimise.

Replaced single table name passing into the select array with a new
variable of arrays called $tables.

Switched LIMIT clause from two seperate keys into one array.

Added 'IS NULL' and 'IS NOT NULL' options to the WHERE clause. New
whereElementNull() function to accomodate this.

Amended whereClause() in /src/classes/DB.php to accept the new $tables
array instead of the previous $conditions array.

Amended whereClause() in /src/classes/DB.php to iterate the tables and
build the where clause from there.

Amended whereElementStandard (), whereElementIn() and
whereElementBetween() to include backticks around table and column
names.

Added a print_r of the query to the showData() function.

Added a template and navigation menu just to improve the examples and
documentation.

Improved commenting.

Added instruction files that can be viewed without connecting to the
database.

Better formatting and mor information in the examples.